refactor: add controllers folders
