+++ app/Http/Controllers/Airline/IndexAirlineController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Airline;
 
 use App\Models\Airline;
 use App\Models\City;
