+++ app/Http/Controllers/City/IndexCityController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\City;
 
 use App\Models\City;
 use Illuminate\Contracts\View\View;
